bots/trading/requirements.txt
python-socketio[client]==5.11.0
requests==2.31.0
coinbase-advanced-py>=1.2.0
python-dotenv>=1.0.0